requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.5.0
pillow==10.4.0
gunicorn==23.0.0
python-dotenv==1.0.1
requests==2.32.3
sentence-transformers==3.1.1
scikit-learn==1.5.2
numpy==2.1.1
PyPDF2==3.0.1
python-multipart==0.0.9
email-validator==2.2.0
gTTS==2.5.3
groq==0.9.0
torch==2.5.1
pypdf==5.1.0
